+++ core/types/header_ext_test.go
@@ -8,7 +8,9 @@ import (
 	"encoding/json"
 	"math/big"
 	"reflect"
+	"slices"
 	"testing"
+	"unsafe"
 
 	"github.com/ava-labs/libevm/common"
 	ethtypes "github.com/ava-labs/libevm/core/types"
@@ -73,13 +75,13 @@ func TestHeaderWithNonZeroFields(t *testing.T) {
 	t.Parallel()
 
 	header, extra := headerWithNonZeroFields()
-	t.Run("Header", func(t *testing.T) { allExportedFieldsSet(t, header) })
-	t.Run("HeaderExtra", func(t *testing.T) { allExportedFieldsSet(t, extra) })
+	t.Run("Header", func(t *testing.T) { allFieldsSet(t, header, "extra") })
+	t.Run("HeaderExtra", func(t *testing.T) { allFieldsSet(t, extra) })
 }
 
 // headerWithNonZeroFields returns a [Header] and a [HeaderExtra],
 // each with all fields set to non-zero values.
-// The [HeaderExtra] extra payload is set in the [Header] via [SetHeaderExtra].
+// The [HeaderExtra] extra payload is set in the [Header] via [WithHeaderExtra].
 //
 // NOTE: They can be used to demonstrate that RLP and JSON round-trip encoding
 // can recover all fields, but not that the encoded format is correct. This is
@@ -112,26 +114,34 @@ func headerWithNonZeroFields() (*Header, *HeaderExtra) {
 		ExtDataGasUsed: big.NewInt(22),
 		BlockGasCost:   big.NewInt(23),
 	}
-	SetHeaderExtra(header, extra)
-	return header, extra
+	return WithHeaderExtra(header, extra), extra
 }
 
-func allExportedFieldsSet[T interface {
-	ethtypes.Header | HeaderExtra
-}](t *testing.T, x *T) {
+func allFieldsSet[T interface {
+	Header | HeaderExtra | Block | Body | BlockBodyExtra
+}](t *testing.T, x *T, ignoredFields ...string) {
 	// We don't test for nil pointers because we're only confirming that
 	// test-input data is well-formed. A panic due to a dereference will be
 	// reported anyway.
 
-	v := reflect.ValueOf(*x)
+	v := reflect.ValueOf(x).Elem()
 	for i := range v.Type().NumField() {
 		field := v.Type().Field(i)
-		if !field.IsExported() {
+		if slices.Contains(ignoredFields, field.Name) {
 			continue
 		}
 
 		t.Run(field.Name, func(t *testing.T) {
-			switch f := v.Field(i).Interface().(type) {
+			fieldValue := v.Field(i)
+			if !field.IsExported() {
+				// Note: we need to check unexported fields especially for [Block].
+				if fieldValue.Kind() == reflect.Ptr {
+					require.Falsef(t, fieldValue.IsNil(), "field %q is nil", field.Name)
+				}
+				fieldValue = reflect.NewAt(fieldValue.Type(), unsafe.Pointer(fieldValue.UnsafeAddr())).Elem() //nolint:gosec
+			}
+
+			switch f := fieldValue.Interface().(type) {
 			case common.Hash:
 				assertNonZero(t, f)
 			case common.Address:
@@ -139,6 +149,8 @@ func allExportedFieldsSet[T interface {
 			case BlockNonce:
 				assertNonZero(t, f)
 			case Bloom:
+				assertNonZero(t, f)
+			case uint32:
 				assertNonZero(t, f)
 			case uint64:
 				assertNonZero(t, f)
@@ -148,7 +160,11 @@ func allExportedFieldsSet[T interface {
 				assertNonZero(t, f)
 			case *uint64:
 				assertNonZero(t, f)
-			case []uint8:
+			case *[]uint8:
+				assertNonZero(t, f)
+			case *Header:
+				assertNonZero(t, f)
+			case []uint8, []*Header, Transactions, []*Transaction, ethtypes.Withdrawals, []*ethtypes.Withdrawal:
 				assert.NotEmpty(t, f)
 			default:
 				t.Errorf("Field %q has unsupported type %T", field.Name, f)
@@ -158,8 +174,8 @@ func allExportedFieldsSet[T interface {
 }
 
 func assertNonZero[T interface {
-	common.Hash | common.Address | BlockNonce | uint64 | Bloom |
-		*big.Int | *common.Hash | *uint64
+	common.Hash | common.Address | BlockNonce | uint32 | uint64 | Bloom |
+		*big.Int | *common.Hash | *uint64 | *[]uint8 | *Header
 }](t *testing.T, v T) {
 	t.Helper()
 	var zero T
@@ -167,5 +183,7 @@ func assertNonZero[T interface {
 		t.Errorf("must not be zero value for %T", v)
 	}
 }
+
+// Note [TestCopyHeader] tests the [HeaderExtra.PostCopy] method.
 
 func ptrTo[T any](x T) *T { return &x }